+++ client/src/features/ExpandedNote/ui/Controls.tsx
