+++ client/src/features/ExpandedNote/ui/Header.tsx
